Move methods of ParserService to MapperService Delete ParserService Add UpdateGenerator to test folder Add Tests for MapperService
