+++ src/test/java/pro/sky/animalshelter4/listener/TelegramBotUpdatesListenerTest.java
@@ -16,6 +16,7 @@ import org.mockito.*;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
+import pro.sky.animalshelter4.UpdateGenerator;
 import pro.sky.animalshelter4.model.Command;
 import pro.sky.animalshelter4.entity.Chat;
 import pro.sky.animalshelter4.entity.City;
@@ -35,14 +36,13 @@ import static org.mockito.Mockito.*;
 @ExtendWith(MockitoExtension.class)
 class TelegramBotUpdatesListenerTest {
     private final TelegramBot telegramBot = mock(TelegramBot.class);
-    private final ParserService parserService = new ParserService();
-    private final MapperService mapperService = new MapperService(parserService);
+    private final MapperService mapperService = new MapperService();
     private final TelegramBotSenderService telegramBotSenderService = new TelegramBotSenderService(telegramBot);
     private final TelegramBotContentSaver telegramBotContentSaver = new TelegramBotContentSaver("./materials", telegramBotSenderService, telegramBot);
     private final TelegramBotUpdatesService telegramBotUpdatesService = new TelegramBotUpdatesService(telegramBotSenderService, mapperService, telegramBotContentSaver);
     private final TelegramBotUpdatesListener telegramBotUpdatesListener = new TelegramBotUpdatesListener(telegramBot, telegramBotUpdatesService);
-    ;
-    private final Faker faker = new Faker();
+
+    private final UpdateGenerator updateGenerator = new UpdateGenerator();
 
     @BeforeEach
     public void init() {
@@ -53,7 +53,7 @@ class TelegramBotUpdatesListenerTest {
         Long id = 50L;
         String command = Command.START.getTitle();
         List<Update> updateList = new ArrayList<>(List.of(
-                generateUpdateMessageWithReflection("", "", "", id, command)));
+                updateGenerator.generateUpdateMessageWithReflection("", "", "", id, command)));
         telegramBotUpdatesListener.process(updateList);
         ArgumentCaptor<SendMessage> argumentCaptor = ArgumentCaptor.forClass(SendMessage.class);
         Mockito.verify(telegramBot, times(2)).execute(argumentCaptor.capture());
@@ -73,9 +73,9 @@ class TelegramBotUpdatesListenerTest {
         Long id = 50L;
         String command = Command.INFO.getTitle();
         List<Update> updateList = new ArrayList<>(List.of(
-                generateUpdateMessageWithReflection("", "", "", id, command),
-                generateUpdateMessageWithReflection("", "", "", id, command + " 1"),
-                generateUpdateCallbackQueryWithReflection("", "", "", id, command)));
+                updateGenerator.generateUpdateMessageWithReflection("", "", "", id, command),
+                updateGenerator.generateUpdateMessageWithReflection("", "", "", id, command + " 1"),
+                updateGenerator.generateUpdateCallbackQueryWithReflection("", "", "", id, command)));
         telegramBotUpdatesListener.process(updateList);
         ArgumentCaptor<SendMessage> argumentCaptor = ArgumentCaptor.forClass(SendMessage.class);
         Mockito.verify(telegramBot, times(6)).execute(argumentCaptor.capture());
@@ -106,8 +106,8 @@ class TelegramBotUpdatesListenerTest {
         Long id = 50L;
         String command = Command.HOW.getTitle();
         List<Update> updateList = new ArrayList<>(List.of(
-                generateUpdateMessageWithReflection("", "", "", id, command),
-                generateUpdateCallbackQueryWithReflection("", "", "", id, command)));
+                updateGenerator.generateUpdateMessageWithReflection("", "", "", id, command),
+                updateGenerator.generateUpdateCallbackQueryWithReflection("", "", "", id, command)));
         telegramBotUpdatesListener.process(updateList);
         ArgumentCaptor<SendMessage> argumentCaptor = ArgumentCaptor.forClass(SendMessage.class);
         Mockito.verify(telegramBot, times(4)).execute(argumentCaptor.capture());
@@ -132,7 +132,7 @@ class TelegramBotUpdatesListenerTest {
         Long id = 50L;
         String command = "/fegfdhesfhdgmghrfdgg";
         List<Update> updateList = new ArrayList<>(List.of(
-                generateUpdateMessageWithReflection("", "", "", id, command)));
+                updateGenerator.generateUpdateMessageWithReflection("", "", "", id, command)));
         telegramBotUpdatesListener.process(updateList);
         ArgumentCaptor<SendMessage> argumentCaptor = ArgumentCaptor.forClass(SendMessage.class);
         Mockito.verify(telegramBot, times(2)).execute(argumentCaptor.capture());
@@ -151,7 +151,7 @@ class TelegramBotUpdatesListenerTest {
         Long id = 50L;
         String command = "fegfdhesfhdgmghrfdgg";
         List<Update> updateList = new ArrayList<>(List.of(
-                generateUpdateMessageWithReflection("", "", "", id, command)));
+                updateGenerator.generateUpdateMessageWithReflection("", "", "", id, command)));
         telegramBotUpdatesListener.process(updateList);
         ArgumentCaptor<SendMessage> argumentCaptor = ArgumentCaptor.forClass(SendMessage.class);
         Mockito.verify(telegramBot, times(2)).execute(argumentCaptor.capture());
@@ -170,187 +170,9 @@ class TelegramBotUpdatesListenerTest {
         Long id = 50L;
         String command = Command.EMPTY_CALLBACK_DATA_FOR_BUTTON.getTitle();
         List<Update> updateList = new ArrayList<>(List.of(
-                generateUpdateCallbackQueryWithReflection("", "", "", id, command)));
+                updateGenerator.generateUpdateCallbackQueryWithReflection("", "", "", id, command)));
         telegramBotUpdatesListener.process(updateList);
         Mockito.verify(telegramBot, times(0)).execute(any());
-    }
-
-
-    private Chat mapUpdateToChat(Update update) {
-        Chat chat = new Chat();
-        if (update.message() != null) {
-            chat.setId(update.message().chat().id());
-            chat.setName(generateNameIfEmpty(update.message().chat().firstName()));
-        } else if (update.callbackQuery() != null) {
-            chat.setId(update.callbackQuery().from().id());
-            chat.setName(generateNameIfEmpty(update.callbackQuery().from().firstName()));
-        }
-        chat.setVolunteer(false);
-        chat.setPhone(generatePhoneIfEmpty(""));
-        City city = new City();
-        city.setCityName(generateCityIfEmpty(""));
-        city.setApproved(true);
-        city.setTimeZone(generateTimeZoneIfNull(null));
-        city.setId(generateIdIfEmpty(null));
-        chat.setCity(city);
-        return chat;
-    }
-
-    private Update generateUpdateCallbackQueryWithReflection(String username,
-                                                             String firstName,
-                                                             String lastName,
-                                                             Long chatId,
-                                                             String callbackQueryData) {
-        username = generateNameIfEmpty(username);
-        firstName = generateNameIfEmpty(firstName);
-        lastName = generateNameIfEmpty(lastName);
-        chatId = generateIdIfEmpty(chatId);
-        callbackQueryData = generateMessageIfEmpty(callbackQueryData);
-
-        Update update = new Update();
-        CallbackQuery callbackQuery = new CallbackQuery();
-        User user = new User(0L);
-
-        try {
-            Field userNameField = user.getClass().getDeclaredField("username");
-            userNameField.setAccessible(true);
-            Field firstNameField = user.getClass().getDeclaredField("first_name");
-            firstNameField.setAccessible(true);
-            Field lastNameField = user.getClass().getDeclaredField("last_name");
-            lastNameField.setAccessible(true);
-            Field userId = user.getClass().getDeclaredField("id");
-            userId.setAccessible(true);
-            userNameField.set(user, username);
-            firstNameField.set(user, firstName);
-            lastNameField.set(user, lastName);
-            userId.set(user, chatId);
-
-            Field callbackUserField = callbackQuery.getClass().getDeclaredField("from");
-            callbackUserField.setAccessible(true);
-            Field callbackDataField = callbackQuery.getClass().getDeclaredField("data");
-            callbackDataField.setAccessible(true);
-            callbackUserField.set(callbackQuery, user);
-            callbackDataField.set(callbackQuery, callbackQueryData);
-
-            Field updateCallbackField = update.getClass().getDeclaredField("callback_query");
-            updateCallbackField.setAccessible(true);
-            updateCallbackField.set(update, callbackQuery);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            e.printStackTrace();
-        }
-        return update;
-    }
-
-    private Update generateUpdateMessageWithReflection() {
-        return generateUpdateMessageWithReflection("", "", "", -1L, "");
-    }
-
-    private Update generateUpdateMessageWithReflection(String username,
-                                                       String firstName,
-                                                       String lastName,
-                                                       Long chatId,
-                                                       String messageText) {
-        username = generateNameIfEmpty(username);
-        firstName = generateNameIfEmpty(firstName);
-        lastName = generateNameIfEmpty(lastName);
-        messageText = generateMessageIfEmpty(messageText);
-        chatId = generateIdIfEmpty(chatId);
-
-        Update update = new Update();
-        Message message = new Message();
-        com.pengrad.telegrambot.model.Chat chat = new com.pengrad.telegrambot.model.Chat();
-        User user = new User(0L);
-
-        try {
-            Field userNameField = user.getClass().getDeclaredField("username");
-            userNameField.setAccessible(true);
-            Field firstNameField = user.getClass().getDeclaredField("first_name");
-            firstNameField.setAccessible(true);
-            Field lastNameField = user.getClass().getDeclaredField("last_name");
-            lastNameField.setAccessible(true);
-            Field userId = user.getClass().getDeclaredField("id");
-            userId.setAccessible(true);
-            userNameField.set(user, username);
-            firstNameField.set(user, firstName);
-            lastNameField.set(user, lastName);
-            userId.set(user, chatId);
-
-
-            Field chatIdField = chat.getClass().getDeclaredField("id");
-            chatIdField.setAccessible(true);
-            chatIdField.set(chat, chatId);
-
-            Field messageTextField = message.getClass().getDeclaredField("text");
-            messageTextField.setAccessible(true);
-            Field messageChatField = message.getClass().getDeclaredField("chat");
-            messageChatField.setAccessible(true);
-            Field messageUserField = message.getClass().getDeclaredField("from");
-            messageUserField.setAccessible(true);
-            messageTextField.set(message, messageText);
-            messageChatField.set(message, chat);
-            messageUserField.set(message, user);
-
-            Field updateMessageField = update.getClass().getDeclaredField("message");
-            updateMessageField.setAccessible(true);
-            updateMessageField.set(update, message);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            e.printStackTrace();
-        }
-        return update;
-    }
-
-    private int generateTimeZoneIfNull(Integer timeZone) {
-        if (timeZone == null || timeZone < 0) {
-            timeZone = faker.random().nextInt(-11, 12);
-        }
-        return timeZone;
-    }
-
-    private String generateAddressIfEmpty(String address) {
-        if (address == null || address.length() == 0) {
-            return faker.address().streetAddress();
-        }
-        return address;
-    }
-
-    private String generateCityIfEmpty(String city) {
-        if (city == null || city.length() == 0) {
-            return faker.address().city();
-        }
-        return city;
-    }
-
-    private String generatePhoneIfEmpty(String phone) {
-        if (phone == null || phone.length() == 0) {
-            return faker.phoneNumber().phoneNumber();
-        }
-        return phone;
-    }
-
-    private String generateNameIfEmpty(String name) {
-        if (name == null || name.length() == 0) {
-            return faker.name().username();
-        }
-        return name;
-    }
-
-    private Long generateIdIfEmpty(Long id) {
-        if (id == null || id < 0) {
-            long idTemp = -1L;
-            //id with <100 I leave for my needs
-            while (idTemp < 100) {
-                idTemp = faker.random().nextLong();
-            }
-            return idTemp;
-        }
-        return id;
-    }
-
-    private String generateMessageIfEmpty(String message) {
-        if (message == null || message.length() == 0) {
-            return faker.lordOfTheRings().character();
-        }
-        return message;
     }
 
 }
